Optimize calculation with memoization
There are only so many board positions to evaluate (3**9 = 19683, and
some of those are even invalid), and we revisit most of those. Add a
memoization decorator for the score function, so we don't have to
recalculate it.

On my machine, this lowers the runtime from 5.82 seconds to 0.18 seconds
for AI vs AI.